Add FM_EREPORT_TIME when event is posted
Previously I was adding the FM_EREPORT_TIME time when the nvlist
was constructed.  However, with the update to onnv_141 these
ereport nvlists are now constructed in several places and it
doesn't make sense for each of them to have to add this common
bit of info.  To handle this the FM_EREPORT_TIME is now only
added once when the event is posted.