Set zv->zv_objset for zil_replay() case
For the case where we have a zil to replay we need to ensure that
zv->zv_objset contains the current objset.  Since the caller has
a hold on the object set it is safe to pass to zil_replay as part
of the zv.  Call path zvol_create_minor()->zil_replay()->
zil_parse()->zil_replay_log_record()->zvol_replay_write().