Honor spa_config_path for ztest '-f' option
The new spa_config_path string was lost from ztest_run_zdb() during
the onnv_141 merge.  This commit puts it back in place so the '-f'
option is properly honored.

Additionally this function had been tweaked so ztest could be run
in-tree but that broke running it when installed as a package.  I've
updated that chunk to detect where it's running and try to do the
right thing in both cases.

Closes #49