Quiet two more large allocation warnings
Both of these allocations are larger than 8k which trips the
warning, but so no large that they must be vmem_alloc()'ed.  For
this case I have added the KM_NODEBUG flag which silences the
warning.  The flag also acts as a marker in the source so I can
keep track of this as something which needs to be better addressed
in the long term.

SPL: large kmem_alloc(11876, 0x50) at get_nvlist:911 (341009/1877530)
SPL: large kmem_alloc(11240, 0x50) at load_nvlist:1273 (2639492/2643588)

Closes #50