Make interlibrary dependencies explicit
Under the latest Gentoo (2.6.34-gentoo-r2) the interlibrary
dependencies must be made explicit.  This is likely due to
the --as-needed linker option which is commonly used under
Gentoo.  This option requires that the linker only include
the minimum set of symbols required by the binary.  Without
the full dependency tree the linker would need to bring in
all the symbols from dependent libraries.

Signed-off-by: Brian Behlendorf <behlendorf1@llnl.gov>