Cleanly handle partitions in zpool_layout
I knew I'd forget something in the first version of this script.
It turns out what I forgot was partition handling, this commit
ensures duplicate entries are not created in the zdev.conf file.
This previously could occur due to partition entires in the by-path
directory.  This is not needed the existing udev rules will ensure
the partitions are created in /dev/disk/zpool/.