Support udev shorthand paths
The Solaris code allows shorthand paths by checking for the short
name in /dev/dsk/.  For Linux we extend this shorthand syntax to
be udev aware.  This commit now additionally checks for the short
name in /dev/, /dev/disk/by-id, /dev/disk/by-label, /dev/disk/by-path,
/dev/disk/by-uuid, and /dev/disk/zpool/.  If the short name is found
in any of these places the path is fully expanded and used.  This
may allow for the simpler creation of pools:

  # Using /dev/disk/by-id/ short names
  zpool crete tank raidz ata-Maxtor_7Y250M0_Y638DK2E \
      ata-Maxtor_7Y250M0_Y638PBNE ata-Maxtor_7Y250M0_Y638PBZE

  # Using /dev/disk/zpool/ short names
  zpool create tank raidz2 A1 A2 A3 A4 mirror B1 B2 B3 B4

  # Using full path names
  zpool create tank raidz2 /dev/sd[ah]